+++ src/beatdrop/__init__.py
@@ -1,6 +1,6 @@
 
 
-__version__ = "0.1.0a5"
+__version__ = "0.1.0a7"
 __all__ = [
     "art",
     "exceptions"