+++ src/beatdrop/schedulers/rq_redis_scheduler.py
@@ -2,25 +2,20 @@
 
 from pydantic.dataclasses import dataclass
 
-from beatdrop.schedulers.celery_scheduler import CeleryScheduler
-from beatdrop.schedulers.sql_scheduler import SQLScheduler
+from beatdrop.schedulers.rq_scheduler import RQScheduler
+from beatdrop.schedulers.redis_scheduler import RedisScheduler
 
 
 class Config:
     arbitrary_types_allowed = True
 
 @dataclass(config=Config)
-class CelerySQLScheduler(SQLScheduler, CeleryScheduler):
-    """Hold schedule entries in an SQL database, and send tasks to Celery queues.
+class RQRedisScheduler(RedisScheduler, RQScheduler):
+    """Hold schedule entries in Redis, and send to RQ (Redis Queue) task queues.
 
-    Uses an SQL database to store schedule entries and scheduler state.
-    It is safe to run multiple ``CelerySQLScheduler`` s simultaneously, 
-    as well as have many that are purely used as clients to read/write entries.
-
-    **NOTE** - You must also install the DB driver specified in the URL for ``create_engine_kwargs``.
-
-    **NOTE** - Before running the scheduler for the first time the 
-    DB tables must be created using ``create_tables()``.
+    Uses Redis to store schedule entries and scheduler state.
+    It is safe to run multiple ``RQRedisScheduler`` s simultaneously, 
+    as well as have many that are used as clients to read/write entries.
 
     Parameters
     ----------
@@ -34,30 +29,32 @@ class CelerySQLScheduler(SQLScheduler, CeleryScheduler):
         In general these entries are not held in non-volatile storage 
         so any metadata they hold will be lost if the scheduler fails.
         These entries are static.  The keys cannot be overwritten or deleted.
-    lock_timeout: datetime.timedelta
+    lock_timeout : datetime.timedelta
         The time a scheduler does not refresh the scheduler lock before it is considered dead. 
         Should be at least 3 times the ``max_interval``.
-    create_engine_kwargs: dict
-        Keyword arguments to pass to ``sqlalchemy.create_engine``.
-        See SQLAlchemy docs for more info. 
-        https://docs.sqlalchemy.org/en/14/core/engines.html#sqlalchemy.create_engine
-    celery_app : celery.Celery
-        Celery app for sending tasks.
+    redis_py_kwargs : Dict[str, Any]
+        redis-py's ``redis.Redis()`` key word arguments. Some of the client configuration items may be overwritten.
+        https://redis-py.readthedocs.io/en/stable/connections.html#generic-client
+    rq_queue : rq.Queue
+        RQ Queue to send tasks to.
 
     Example
     -------
     .. code-block:: python
 
-        from celery import Celery
-        from beatdrop import CelerySQLScheduler
+        from rq import Queue
+        from beatdrop import RQRedisScheduler
 
-        celery_app = Celery()
-        sched = CelerySQLScheduler(
+        rq_queue = Queue()
+        sched = RQRedisScheduler(
             max_interval=60,
-            celery_app=celery_app,
+            rq_queue=rq_queue,
             lock_timeout=180,
-            create_engine_kwargs={
-                "url": "sqlite:///my_sqlite.db"
+            redis_py_kwargs={
+                "host": "my.redis.host",
+                "port": 6379,
+                "db": 0,
+                "password": "mys3cr3t"
             }
         )
         # use the scheduler as a client
@@ -68,6 +65,3 @@ class CelerySQLScheduler(SQLScheduler, CeleryScheduler):
         # or run it
         sched.run()
     """
-
-
-