+++ tests/unit/celery_scheduler_test.py
@@ -7,14 +7,8 @@ import celery
 import pytest
 import redislite
 
-from beatdrop import messages
 from beatdrop.entries import IntervalEntry
 from beatdrop.schedulers import CeleryScheduler
-
-
-@pytest.fixture(scope="function")
-def celery_app(rdb: redislite.Redis) -> celery.Celery:
-    return celery.Celery("tester", broker="redis+socket://{}?db=0".format(rdb.socket_file))
 
 
 @pytest.fixture